Restructured customerSchema

+++ server/routes/Users.js
@@ -2,12 +2,10 @@ const express = require('express');
 const router = express.Router();
 const { User } = require('../models/users');
 const jwt = require('jsonwebtoken');
-const { auth } = require('../middleware/auth');
+const mongoose = require('mongoose');
 
 router
   .get('/', (req, res) => {
-    const { userId, isAdmin } = req.user;
-
     User.find()
       .sort({ createdAt: -1 })
       .then((result) => {
@@ -25,9 +23,14 @@ router
   })
   .get('/singleuser/:id', (req, res) => {
     const { isAdmin } = req.user;
+    const userId = req.params.id;
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ error: 'Ogiltigt användar-ID.' });
+    }
 
     if (isAdmin) {
-      User.findById(req.params.id)
+      User.findById(userId)
         .then((result) => {
           res.json(result);
         })
@@ -36,43 +39,49 @@ router
         });
     }
   })
-  .put('/singleuser/:id', async (req, res) => {
+  .patch('/singleuser/:id', async (req, res) => {
     const userId = req.params.id;
-    const { isAdmin } = req.user;
+    const { isAdmin, userId: loggedInUser } = req.user;
 
-    if (isAdmin) {
-      try {
-        const existingUser = await User.findById(userId);
-        if (!existingUser) {
-          return res.status(404).json({ error: 'Användaren hittades inte.' });
-        }
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ error: 'Ogiltigt användar-ID.' });
+    }
 
-        const newName = req.body.name;
-        const newEmail = req.body.email;
-        const newPassword = req.body.password;
-        const newRole = req.body.isAdmin;
-
-        newName && (existingUser.name = newName);
-        newEmail && (existingUser.email = newEmail);
-        newPassword && (existingUser.password = newPassword);
-        newRole !== undefined && (existingUser.isAdmin = newRole);
-
-        const updatedUser = await existingUser.save();
-        res.status(200).json(updatedUser);
-      } catch (error) {
-        res.status(500).json({ error: 'Ett fel inträffade vid uppdatering av användaren.' });
+    if (!isAdmin && userId !== loggedInUser) {
+      return res.status(403).json({ error: 'Obehörig åtkomst till att ändra vald användare.' });
+    }
+    try {
+      const existingUser = await User.findById(userId);
+      if (!existingUser) {
+        return res.status(404).json({ error: 'Användaren hittades inte.' });
       }
+
+      const { name, email, password, isAdmin: newRole } = req.body.newData;
+
+      name && (existingUser.name = name);
+      email && (existingUser.email = email);
+      password && (existingUser.password = password);
+      newRole !== undefined && (existingUser.isAdmin = newRole);
+
+      const updatedUser = await existingUser.save();
+      res.status(200).json(updatedUser);
+    } catch (error) {
+      res.status(500).json({ error: 'Ett fel inträffade vid uppdatering av användaren.' });
     }
   })
   .delete('/singleuser/:id', async (req, res) => {
     const userId = req.params.id;
     const { isAdmin } = req.user;
 
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ error: 'Ogiltigt användar-ID.' });
+    }
+
     if (isAdmin) {
       try {
         await User.findByIdAndDelete(userId);
 
-        res.status(200).json({ success: 'Användare borttagen' });
+        res.status(200).send('Användare borttagen.');
       } catch (error) {
         console.error('Error deleting user:', error);
         res.status(500).json({ error: 'Ett fel inträffade vid borttagning av användare.' });
@@ -108,8 +117,8 @@ router
       }
       const user = new User({ ...req.body });
       const result = await user.save();
-
-      res.status(201).json({ success: true, user: result });
+      console.log(result);
+      res.status(201).send(result);
     } catch (err) {
       console.error(err);
       res.status(500).json({ error: 'Internal Server Error' });
